Данные с сервера перенесены в глобальную переменную внутри файла getDataFromServer
